+++ components/ui/button/Button.vue
@@ -19,7 +19,7 @@ const props = withDefaults(defineProps<Props>(), {
   <Primitive
     :as="as"
     :as-child="asChild"
-    :class="cn(buttonVariants({ variant, size }), props.class)"
+    :class="cn('btn', buttonVariants({ variant, size }), props.class)"
   >
     <slot />
   </Primitive>
