+++ package.json
@@ -25,7 +25,7 @@
         "vue-router": "latest"
     },
     "devDependencies": {
-        "prez-ui": "^4.0.0",
+        "prez-ui": "^4.0.1",
         "typescript": "^5.6.3"
     }
 }
